# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Directorio de empresas - CCGAMT",
        "description": "Backend y API Acceso a datos.",
        "contact": {
            "email": "brayan-diaz3@upc.edu.co"
        },
        "license": {
            "name": "@Universidad Piloto de Colombia SAM - Todos los Derechos Reservados"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/rachas": {
            "get": {
                "tags": [
                    "Rachas"
                ],
                "summary": "Listar todas las rachas",
                "operationId": "f77cce699c05b9a114ee003b9d0cffa3",
                "responses": {
                    "200": {
                        "description": "Lista de rachas"
                    }
                }
            },
            "post": {
                "tags": [
                    "Rachas"
                ],
                "summary": "Crear una nueva racha",
                "operationId": "2d068205842f86720bd1d0e3469886e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_usuario",
                                    "dias_consecutivos"
                                ],
                                "properties": {
                                    "id_usuario": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "dias_consecutivos": {
                                        "type": "integer",
                                        "example": 5
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Racha creada"
                    },
                    "400": {
                        "description": "Datos inválidos"
                    }
                }
            }
        },
        "/api/rachas/{id}": {
            "get": {
                "tags": [
                    "Rachas"
                ],
                "summary": "Obtener una racha por ID",
                "operationId": "8dc925ecca7151ea7e4bc662263064dc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la racha",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Racha encontrada"
                    },
                    "404": {
                        "description": "Racha no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Rachas"
                ],
                "summary": "Actualizar una racha existente",
                "operationId": "2ade8c680ef3bdcc27b46a897c297ca0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la racha",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "dias_consecutivos"
                                ],
                                "properties": {
                                    "dias_consecutivos": {
                                        "type": "integer",
                                        "example": 10
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Racha actualizada"
                    },
                    "400": {
                        "description": "Datos inválidos"
                    },
                    "404": {
                        "description": "Racha no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Rachas"
                ],
                "summary": "Eliminar una racha",
                "operationId": "af1b5ea784c8b11c102fb3ccbef6fc6b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la racha",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Racha eliminada correctamente"
                    },
                    "404": {
                        "description": "Racha no encontrada"
                    }
                }
            }
        },
        "/api/reportes": {
            "get": {
                "tags": [
                    "Reportes"
                ],
                "summary": "Listar todos los reportes",
                "operationId": "f942c2f53a8662a6f875d951eb2c85b9",
                "responses": {
                    "200": {
                        "description": "Lista de reportes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Reportes"
                ],
                "summary": "Crear un nuevo reporte",
                "operationId": "e19cc71ac5aa411bd87112025d2d8af8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_usuario",
                                    "descripcion"
                                ],
                                "properties": {
                                    "id_usuario": {
                                        "type": "integer"
                                    },
                                    "descripcion": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Reporte creado"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/reportes/{id}": {
            "get": {
                "tags": [
                    "Reportes"
                ],
                "summary": "Obtener un reporte por ID",
                "operationId": "25653ad638314a9e4580de79483d7633",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reporte encontrado"
                    },
                    "404": {
                        "description": "Reporte no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Reportes"
                ],
                "summary": "Actualizar un reporte existente",
                "operationId": "d3d32f5bfae9113e0fe7c5a3ba5485a1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "descripcion": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reporte actualizado"
                    },
                    "400": {
                        "description": "Error de validación"
                    },
                    "404": {
                        "description": "Reporte no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Reportes"
                ],
                "summary": "Eliminar un reporte",
                "operationId": "2db54d0fea33418092a1fb41056e9ec2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Reporte eliminado"
                    },
                    "404": {
                        "description": "Reporte no encontrado"
                    }
                }
            }
        },
...
    },
...
}